recommendation system/P9_01_scripts/az function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-cosmos
pandas==1.1.3
numpy==1.20.0
scikit-learn==0.23.2
scikit-surprise==1.1.1